+++ CHANGELOG.md
@@ -18,6 +18,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Remove demos from the documentation ([#452](https://github.com/studiometa/ui/pull/452), [954cb04](https://github.com/studiometa/ui/commit/954cb04))
 
+### Dependencies
+
+- Update dependency @studiometa/js-toolkit to v3.2.0 ([#457](https://github.com/studiometa/ui/pull/457), [a310b09](https://github.com/studiometa/ui/commit/a310b09))
+
 ## [v1.2.0](https://github.com/studiometa/ui/compare/1.1.1..1.2.0) (2025-08-20)
 
 ### Added